Common Point trait
